// src/data/Interfaces/Rouanet.ts
interface Rouanet {
    id_projeto: 57030;
    pronac: string;
    ano_projeto: string;
    nome: string;
    segmento: string;
    area: string;
    uf: string;
    municipio: string;
    data_inicio: string;
    data_termino: string;
    situacao: string;
    mecanismo: string;
    enquadramento: string;
    valor_captado: string;
    valor_aprovado: string;
    acessibilidade: string;
    objetivo?: string;
    justificativa: string;
    etapa: string;
    ficha_tecnica: string;
    impacto_ambiental: string;
    especificacao_tecnica: string;
}

export default Rouanet;